BellmanFord Algorithm for negative weight edges.
Dijikstra's algorithm cannot deal with negative weight edges.So , We can use Bellman Ford Algorithm to find the minimum distance to all the nodes from the source node. 
DrawBack of this algorithm is that it cannot find the solution when there is a negative weight cycle(sum of weights in a cycle is negative).